extract contractdealproposal types
